LongpressGesture on CreateMemeController shows only the top menu bar.  So we now have the ability to sent previously sent memes.
This state is more stable than the one obtained in earlier versions that has a present() leading to another present() which then led to an error of "Attempt to present ViewController whose view is not in the windows hierarchy".  They were two potential resolutions.  Obtaining the root view & then presenting over there.  Or avoiding a present into another present.  Second option was chosen here.